refactor: push simple functions up to abstract base class.
